CHG: drawRotatedSquares()
1. translateBy() translates (moves) the current transformation matrix.
2. rotate(by:) rotates the current transformation matrix.
3. strokePath() strokes the path with your specified line width, which is 1 if you don't set it explicitly.

In UIKit, you rotate drawing around the center of your view, as if a pin was stuck right through the middle. In Core Graphics, you rotate around the top-left corner, so to avoid that we're going to move the transformation matrix half way into our image first so that we've effectively moved the rotation point.

This also means we need to draw our rotated squares so they are centered on our center: for example, setting their top and left coordinates to be -128 and their width and height to be 256.